requirements-dev.txt
django-debug-toolbar==4.3.0

mypy==1.8.0
isort==5.13.2

# Types
django-stubs==4.2.7
